requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
python-coveralls==2.9.1
#coverage==4.0.3
decorator==4.4.0
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
jedi==0.13.3
numpy==1.16.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.9.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
Pygments==2.3.1
pytest==4.4.1
pytest-cov==2.6.1
pytest-cover==3.0.0
pytest-coverage==0.0
python-dotenv==0.10.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
smmap2==2.0.5
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
setuptools==41.0.1

